+++ napari/_qt/qt_viewer.py
@@ -12,6 +12,8 @@ from qtpy.QtCore import QCoreApplication, QObject, Qt
 from qtpy.QtGui import QCursor, QGuiApplication
 from qtpy.QtWidgets import QFileDialog, QSplitter, QVBoxLayout, QWidget
 
+from napari_builtins.io._write import imsave_extensions
+
 from ..components._interaction_box_mouse_bindings import (
     InteractionBoxMouseBindings,
 )
@@ -20,6 +22,7 @@ from ..components.layerlist import LayerList
 from ..errors import MultipleReaderError, ReaderPluginError
 from ..layers.base.base import Layer
 from ..plugins import _npe2
+from ..settings import get_settings
 from ..utils import config, perf
 from ..utils._proxies import ReadOnlyWrapper
 from ..utils.action_manager import action_manager
@@ -64,11 +67,9 @@ from .._vispy import (  # isort:skip
 
 
 if TYPE_CHECKING:
-    from ..components import ViewerModel
     from npe2.manifest.contributions import WriterContribution
 
-from ..settings import get_settings
-from ..utils.io import imsave_extensions
+    from ..components import ViewerModel
 
 
 def _npe2_decode_selected_filter(
