+++ napari_builtins/_tests/test_builtin_get_writer.py
@@ -1,6 +1,6 @@
 import os
 
-from napari.plugins._builtins import write_layer_data_with_plugins
+from napari_builtins.io import write_layer_data_with_plugins
 
 
 # test_plugin_manager fixture is provided by napari_plugin_engine._testsupport
