+++ napari_builtins/_tests/test_reader_plugins.py
