+++ napari_builtins/io/_write.py
@@ -1,79 +1,95 @@
-"""
-Internal napari hook implementations to be registered by the plugin manager
-"""
+import csv
 import os
 import shutil
-from typing import Any, List, Optional, Sequence, Union
+from typing import Any, List, Optional, Tuple, Union
 
 import numpy as np
 
-from ..types import (
-    FullLayerData,
-    LayerData,
-    ReaderFunction,
-    WriterFunction,
-    image_reader_to_layerdata_reader,
-)
-from ..utils.io import (
-    READER_EXTENSIONS,
-    csv_to_layer_data,
-    imsave,
-    imsave_extensions,
-    magic_imread,
-    write_csv,
-)
-from ..utils.misc import abspath_or_url
+from napari.types import FullLayerData
+from napari.utils.io import imsave
+from napari.utils.misc import abspath_or_url
 
 
-def csv_reader_function(path: Union[str, Sequence[str]]) -> List[LayerData]:
-    if not isinstance(path, str):
-        out: List[LayerData] = []
-        for p in path:
-            layer_data = csv_to_layer_data(p, require_type=None)
-            if layer_data:
-                out.append(layer_data)
-        return out
-    else:
-        layer_data = csv_to_layer_data(path, require_type=None)
-        return [layer_data] if layer_data else []
-
-
-def npy_to_layer_data(path: Union[str, Sequence[str]]) -> List[LayerData]:
-    if isinstance(path, str):
-        return [(np.load(path),)]
-
-    return [(np.load(p),) for p in path]
-
-
-def napari_get_reader(path: Union[str, List[str]]) -> Optional[ReaderFunction]:
-    """Our internal fallback file reader at the end of the reader plugin chain.
-
-    This will assume that the filepath is an image, and will pass all of the
-    necessary information to viewer.add_image().
+def write_csv(
+    filename: str,
+    data: Union[List, np.ndarray],
+    column_names: Optional[List[str]] = None,
+):
+    """Write a csv file.
 
     Parameters
     ----------
-    path : str
-        path to file/directory
+    filename : str
+        Filename for saving csv.
+    data : list or ndarray
+        Table values, contained in a list of lists or an ndarray.
+    column_names : list, optional
+        List of column names for table data.
+    """
+    with open(filename, mode='w', newline='') as csvfile:
+        writer = csv.writer(
+            csvfile,
+            delimiter=',',
+            quotechar='"',
+            quoting=csv.QUOTE_MINIMAL,
+        )
+        if column_names is not None:
+            writer.writerow(column_names)
+        for row in data:
+            writer.writerow(row)
+
+
+def imsave_extensions() -> Tuple[str, ...]:
+    """Valid extensions of files that imsave can write to.
 
     Returns
     -------
-    callable
-        function that returns layer_data to be handed to viewer._add_layer_data
+    tuple
+        Valid extensions of files that imsave can write to.
     """
-    if isinstance(path, str):
-        if path.endswith('.csv'):
-            return csv_reader_function
-        if os.path.isdir(path):
-            return image_reader_to_layerdata_reader(magic_imread)
-        if path.endswith('.npy'):
-            return npy_to_layer_data
-        path = [path]
+    # import imageio
+    # return tuple(set(x for f in imageio.formats for x in f.extensions))
 
-    if all(str(x).lower().endswith(tuple(READER_EXTENSIONS)) for x in path):
-        return image_reader_to_layerdata_reader(magic_imread)
-
-    return None
+    # The above method generates a lot of extensions that will fail.  This list
+    # is a more realistic set, generated by trying to write a variety of numpy
+    # arrays (skimage.data.camera, grass, and some random numpy arrays/shapes).
+    # TODO: maybe write a proper imageio plugin.
+    return (
+        '.bmp',
+        '.bsdf',
+        '.bw',
+        '.eps',
+        '.gif',
+        '.icns',
+        '.ico',
+        '.im',
+        '.j2c',
+        '.j2k',
+        '.jfif',
+        '.jp2',
+        '.jpc',
+        '.jpe',
+        '.jpeg',
+        '.jpf',
+        '.jpg',
+        '.jpx',
+        '.lsm',
+        '.mpo',
+        '.npz',
+        '.pbm',
+        '.pcx',
+        '.pgm',
+        '.png',
+        '.ppm',
+        '.ps',
+        '.rgb',
+        '.rgba',
+        '.sgi',
+        '.stk',
+        '.tga',
+        '.tif',
+        '.tiff',
+    )
 
 
 def napari_write_image(path: str, data: Any, meta: dict) -> Optional[str]:
@@ -162,7 +178,7 @@ def napari_write_points(path: str, data: Any, meta: dict) -> Optional[str]:
     properties = meta.get('properties', {})
     # TODO: we need to change this to the axis names once we get access to them
     # construct table from data
-    column_names = ['axis-' + str(n) for n in range(data.shape[1])]
+    column_names = [f'axis-{str(n)}' for n in range(data.shape[1])]
     if properties:
         column_names += properties.keys()
         prop_table = [
@@ -217,7 +233,7 @@ def napari_write_shapes(path: str, data: Any, meta: dict) -> Optional[str]:
     # TODO: we need to change this to the axis names once we get access to them
     # construct table from data
     n_dimensions = max(s.shape[1] for s in data)
-    column_names = ['axis-' + str(n) for n in range(n_dimensions)]
+    column_names = [f'axis-{str(n)}' for n in range(n_dimensions)]
 
     # add shape id and vertex id of each vertex
     column_names = ['index', 'shape-type', 'vertex-index'] + column_names
@@ -246,34 +262,6 @@ def napari_write_shapes(path: str, data: Any, meta: dict) -> Optional[str]:
     # write table to csv file
     write_csv(path, table, column_names)
     return path
-
-
-def napari_get_writer(
-    path: str, layer_types: List[str]
-) -> Optional[WriterFunction]:
-    """Our internal fallback file writer at the end of the writer plugin chain.
-
-    This will create a new folder from the path and call
-    ``napari_write_<layer>`` for each layer using the ``layer.name`` variable
-    to modify the path such that the layers are written to unique files in the
-    folder. It will use the default builtin writer for each layer type.
-
-    Parameters
-    ----------
-    path : str
-        path to file/directory
-
-    Returns
-    -------
-    callable
-        function that accepts the path and a list of layer_data (where
-        layer_data is ``(data, meta, layer_type)``) and writes each layer.
-    """
-    # normally, a plugin would do some logic here to decide whether it supports
-    # the ``path`` extension and layer_types.  But because this is our builtin
-    # "last resort" implementation, we just immediately hand back the writer
-    # function, and let it throw an exception if it fails.
-    return write_layer_data_with_plugins
 
 
 def write_layer_data_with_plugins(
