+++ setup.cfg
@@ -79,7 +79,7 @@ install_requires =
 
 [options.package_data]
 * = *.pyi
-napari =
+napari_builtins =
     builtins.yaml
 
 # for explanation of %(extra)s syntax see:
@@ -171,7 +171,7 @@ console_scripts =
 pytest11 =
     napari = napari.utils._testsupport
 napari.manifest =
-    napari = napari:builtins.yaml
+    napari_builtins = napari_builtins:builtins.yaml
 
 [flake8]
 # Ignores - https://lintlyci.github.io/Flake8Rules
